Refactor ProductData module: move to ExternalServices.mjs and update import in product-listing.js
